+++ src/EffectiveTLD.php
@@ -4,18 +4,22 @@ declare(strict_types=1);
 
 namespace Pdp;
 
-interface EffectiveTLD extends Host, ExternalDomainName
+interface EffectiveTLD extends Host, DomainNameProvider
 {
-    public const ICANN_DOMAINS = 'ICANN_DOMAINS';
-
-    public const PRIVATE_DOMAINS = 'PRIVATE_DOMAINS';
-
-    public const IANA_DOMAINS = 'IANA_DOMAINS';
-
     /**
      * Tells whether the effective TLD has a matching rule in a Public Suffix List.
      */
     public function isKnown(): bool;
+
+    /**
+     * Tells whether the effective TLD has a matching rule in the IANA Root Zone Database.
+     */
+    public function isIANA(): bool;
+
+    /**
+     * Tells whether the effective TLD has a matching rule in the Public Suffix List.
+     */
+    public function isPublicSuffix(): bool;
 
     /**
      * Tells whether the effective TLD has a matching rule in a Public Suffix List ICANN Section.
@@ -26,9 +30,4 @@ interface EffectiveTLD extends Host, ExternalDomainName
      * Tells whether the effective TLD has a matching rule in a Public Suffix List Private Section.
      */
     public function isPrivate(): bool;
-
-    /**
-     * Tells whether the effective TLD has a matching rule in the IANA Root Zone Database.
-     */
-    public function isIANA(): bool;
 }