+++ src/PublicSuffixList.php
@@ -6,8 +6,12 @@ namespace Pdp;
 
 use JsonSerializable;
 
-interface PublicSuffixList extends DomainResolver, JsonSerializable
+interface PublicSuffixList extends DomainNameResolver, JsonSerializable
 {
+    public const ICANN_DOMAINS = 'ICANN_DOMAINS';
+
+    public const PRIVATE_DOMAINS = 'PRIVATE_DOMAINS';
+
     /**
      * Returns an array representation of the Public Suffix List Rules JSON serializable.
      */